Fix multiple empty cards
Previously, selecting "Add Book" would add an unlimited
amount of empty cards. This is fine, until the
removeInputElements function is called, which only removes
the last card and leaves empty cards that could not be
fixed. Now, a variable which tracks
if an empty card has been made so only one empty card can
appear on the screen at a time.